Embedded/tests/requirements.txt
# Tests Module Dependencies

# Core testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Communication dependencies
paho-mqtt>=1.6.1
websockets>=11.0.0
requests>=2.28.0

# AI and ML dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0

# ROS2 dependencies (if needed)
# rclpy>=0.4.0
# sensor-msgs>=0.1.0
# std-msgs>=0.1.0

# Hardware interface dependencies
# smbus2>=0.4.0  # For I2C communication
# RPi.GPIO>=0.7.0  # For GPIO control (if needed)

# Development dependencies
black>=22.0.0
flake8>=4.0.0
mypy>=0.991

